new sign up form styling complete
