+++ client/src/containers/LoginContainer.js
@@ -1,93 +1,93 @@
-import React, { Component } from "react";
-// import Button from "../components/Form/button";
-// import Input from "../components/Form/input";
+// import React, { Component } from "react";
+// // import Button from "../components/Form/button";
+// // import Input from "../components/Form/input";
 
-const googleImg = require("../components/GoogleBtns/google_signin_buttons/web/1x/btn_google_signin_light_pressed_web.png");
+// const googleImg = require("../components/GoogleBtns/google_signin_buttons/web/1x/btn_google_signin_light_pressed_web.png");
     
-class LoginContainer extends Component {
-    constructor(props) {
-        super(props);
+// class LoginContainer extends Component {
+//     constructor(props) {
+//         super(props);
 
-        this.state = {
-            user: null
-        }
-        this.handleLoginSubmit = this.handleLoginSubmit.bind(this);
-        this.handleClearLogin = this.handleClearLogin.bind(this);
-        this.handleLogin = this.handleLogin.bind(this);
-        this.handleLogout = this.handleLogout.bind(this);
-    }
+//         this.state = {
+//             user: null
+//         }
+//         this.handleLoginSubmit = this.handleLoginSubmit.bind(this);
+//         this.handleClearLogin = this.handleClearLogin.bind(this);
+//         this.handleLogin = this.handleLogin.bind(this);
+//         this.handleLogout = this.handleLogout.bind(this);
+//     }
 
-    componentDidMount() {
+//     componentDidMount() {
 
-    }
+//     }
 
-    handleLoginSubmit(event) {
-        // API.login()
+//     handleLoginSubmit(event) {
+//         // API.login()
      
         
-    }
+//     }
 
-    // loadFriendDetail = () => {
-    //     API.getFriendDetail()
-    //         .then(res => this.setState({ friends: res.data }))
-    //         .catch(err => console.log(err))
-    // } 
+//     // loadFriendDetail = () => {
+//     //     API.getFriendDetail()
+//     //         .then(res => this.setState({ friends: res.data }))
+//     //         .catch(err => console.log(err))
+//     // } 
 
-    handleClearLogin(event) {
-        event.preventDefault();
-        this.setState({
-            User: {
-                username: "",
-                password: ""
-            }
-        })
-    }
+//     handleClearLogin(event) {
+//         event.preventDefault();
+//         this.setState({
+//             User: {
+//                 username: "",
+//                 password: ""
+//             }
+//         })
+//     }
 
-    handleLogout() {
+//     handleLogout() {
 
-      }
+//       }
       
-    handleLogin() {
+//     handleLogin() {
 
-      }
+//       }
 
     
-    render() {
-        return (
-            <div className="container center" style={{ marginTop: 60 }}>
-                {this.state.user ?
-                  <button onClick={this.handleLogout}>Log Out</button>                
-                  :
-                  <button onClick={this.handleLogin}><img src={googleImg} alt="Google Login Button"></img></button>              
-                }
-            {/* <form className="card col s6 z-depth-3" onSubmit={this.handleLoginSubmit}> */}
-                {/* Username Input */}
-                {/* <Input 
-                    inputType = {"text"}
-                    title = {"Username"}
-                    name = {"username"}
-                    value = {this.state.User.username}
-                    handleChange = {this.handleLoginInput}
-                /> */}
-                {/* Password Input */}  
-                {/* <Input 
-                    inputType = {"text"}
-                    title = {"Password"}
-                    name = {"password"}
-                    value = {this.state.User.password}
-                    handleChange = {this.handleLoginInput}
-                />   */}
-                {/* Submit */}
-                {/* <Button 
-                    action = {this.handleLoginSubmit}  
-                    type = {"primary"}
-                    title = {"Submit"} 
-                />     
-            </form> */}
-            </div>
-        );
-    }
+//     render() {
+//         return (
+//             <div className="container center" style={{ marginTop: 60 }}>
+//                 {this.state.user ?
+//                   <button onClick={this.handleLogout}>Log Out</button>                
+//                   :
+//                   <button onClick={this.handleLogin}><img src={googleImg} alt="Google Login Button"></img></button>              
+//                 }
+//             {/* <form className="card col s6 z-depth-3" onSubmit={this.handleLoginSubmit}> */}
+//                 {/* Username Input */}
+//                 {/* <Input 
+//                     inputType = {"text"}
+//                     title = {"Username"}
+//                     name = {"username"}
+//                     value = {this.state.User.username}
+//                     handleChange = {this.handleLoginInput}
+//                 /> */}
+//                 {/* Password Input */}  
+//                 {/* <Input 
+//                     inputType = {"text"}
+//                     title = {"Password"}
+//                     name = {"password"}
+//                     value = {this.state.User.password}
+//                     handleChange = {this.handleLoginInput}
+//                 />   */}
+//                 {/* Submit */}
+//                 {/* <Button 
+//                     action = {this.handleLoginSubmit}  
+//                     type = {"primary"}
+//                     title = {"Submit"} 
+//                 />     
+//             </form> */}
+//             </div>
+//         );
+//     }
 
-}
+// }
 
-export default LoginContainer;
+// export default LoginContainer;
