+++ client/src/pages/OldSignup.js
@@ -1,0 +1,129 @@
+// import React, { Component } from "react";
+// import { Link } from "react-router-dom";
+// import axios from "axios";
+// // import "./styles.css";
+// import FormContainer from "../containers/FormContainer";
+// import KeyboardBackspaceIcon from "@material-ui/icons/KeyboardBackspace";
+
+
+
+// class Signup extends Component {
+//   constructor() {
+//     super();
+//     this.state = {
+//       name: "",
+//       email: "",
+//       password: "",
+//       password2: "",
+//       errors: {}
+//     };
+//   }
+//   onChange = e => {
+//     this.setState({ [e.target.id]: e.target.value });
+//   };
+
+//   onSubmit = e => {
+//     e.preventDefault();
+//     const newUser = {
+//       name: this.state.name,
+//       email: this.state.email,
+//       password: this.state.password,
+//       password2: this.state.password2
+//     };
+//     console.log(newUser);
+//     axios.post('/api/users/register', {
+//       name: this.state.name,
+//       email: this.state.email,
+//       password: this.state.password,
+//       password2: this.state.password2
+//       })
+//     .then((result) => {
+//       console.log("New User result: " + result.data._id);
+//       this.props.history.push('/main')
+//     })
+//     .catch((error) => {
+//       console.log(error.response.data);
+//     })
+//   };
+
+//   render() {
+//     const { errors } = this.state;
+//     return (
+//       <div className="container">
+//         <div className="row">
+//           <div className="col s8 offset-s2">
+//             <Link to="/" className="btn-flat waves-effect"><KeyboardBackspaceIcon /> 
+//               Back to home
+//             </Link>
+//             <div className="col s12" style={{ paddingLeft: "11.250px" }}>
+//               <h4>
+//                 <b>Register</b> below
+//               </h4>
+//               <p className="grey-text text-darken-1">
+//                 Already have an account? <Link to="/">Log in</Link>
+//               </p>
+//             </div>
+//             <form noValidate onSubmit={this.onSubmit}>
+//               <div className="input-field col s12">
+//                 <input
+//                   onChange={this.onChange}
+//                   value={this.state.name}
+//                   error={errors.name}
+//                   id="name"
+//                   type="text"
+//                   placeholder="Name"
+//                 />
+//               </div>
+//               <div className="input-field col s12">
+//                 <input
+//                   onChange={this.onChange}
+//                   value={this.state.email}
+//                   error={errors.email}
+//                   id="email"
+//                   type="email"
+//                   placeholder="Email"
+//                 />
+//               </div>
+//               <div className="input-field col s12">
+//                 <input
+//                   onChange={this.onChange}
+//                   value={this.state.password}
+//                   error={errors.password}
+//                   id="password"
+//                   type="password"
+//                   placeholder="Password"
+//                 />
+//               </div>
+//               <div className="input-field col s12">
+//                 <input
+//                   onChange={this.onChange}
+//                   value={this.state.password2}
+//                   error={errors.password2}
+//                   id="password2"
+//                   type="password"
+//                   placeholder="Confirm Password"
+//                 />
+//               </div>
+//               <div className="col s12" style={{ paddingLeft: "11.250px" }}>
+//                 <button
+//                   style={{
+//                     width: "150px",
+//                     borderRadius: "3px",
+//                     letterSpacing: "1.5px",
+//                     marginTop: "1rem"
+//                   }}
+//                   type="submit"
+//                   className="btn btn-large waves-effect waves-light hoverable blue accent-3"
+//                 >
+//                   Sign up
+//                 </button>
+//               </div>
+//             </form>
+//           </div>
+//         </div>
+//       </div>
+//     );
+//   }
+// }
+
+// export default Signup;
