--- client/src/pages/Test.js
@@ -1,19 +1,0 @@
-import React, { Component } from "react";
-// import "./styles.css";
-import Form from "../components/FormTest/Form";
-// import FullLogo from "../components/Logo"
-import "../App.css";
-
-
-class Test extends Component {
-    render() {
-      return (
-          <div className="container">
-            <Form />
-          </div>
-          )
-      }
-  }
-  
-  export default Test;
-
